pyproject.toml
[project]
name = "autenticacao-html-js"
version = "0.1.0"
description = ""
authors = [
    {name = "RaphaVBueno",email = "rafavbueno@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "fastapi (>=0.128.0,<0.129.0)",
    "uvicorn (>=0.40.0,<0.41.0)",
    "python-dotenv (>=1.2.1,<2.0.0)",
    "python-jose[cryptography] (>=3.5.0,<4.0.0)",
    "bcrypt (>=5.0.0,<6.0.0)",
    "sqlalchemy (>=2.0.45,<3.0.0)",
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
